Dodanie tailwinda oraz poprawa kontekstu do admin panel oraz usuniecie smieci

+++ frontend/my-kiosk-frontend/src/components/CategoriesManager.js
@@ -1,10 +1,15 @@
-// src/components/CategoriesManager.js
-import React, { useEffect, useState } from 'react';
+import React, {useContext, useEffect, useState} from 'react';
 import CategoryForm from './CategoryForm';
+import { useCategoryChange } from '../context/CategoryChangeContext';
+import { AdminContext } from '../context/AdminContext';
 
 const CategoriesManager = () => {
     const [categories, setCategories] = useState([]);
     const [editingCategory, setEditingCategory] = useState(null);
+    const { triggerCategoryChange } = useCategoryChange();
+    const { state } = useContext(AdminContext);
+    const adminKey = state.adminKey;
+
 
     useEffect(() => {
         fetchCategories();
@@ -33,7 +38,7 @@ const CategoriesManager = () => {
                 method: method,
                 headers: {
                     'Content-Type': 'application/json',
-                    'Admin-Secret-Key': 'admin', // Replace with actual admin key
+                    'Admin-Secret-Key': adminKey,
                 },
                 body: JSON.stringify(category),
             });
@@ -41,6 +46,8 @@ const CategoriesManager = () => {
             if (response.ok) {
                 fetchCategories();
                 setEditingCategory(null); // Clear editing state
+                triggerCategoryChange();
+
             } else {
                 console.error('Failed to submit category');
             }
@@ -50,46 +57,51 @@ const CategoriesManager = () => {
     };
 
     const deleteCategory = async (id) => {
-        try {
-            const response = await fetch(`http://localhost:3001/api/admin/categories/${id}`, {
-                method: 'DELETE',
-                headers: {
-                    'Admin-Secret-Key': 'admin', // Replace with actual admin key
-                },
-            });
+        if (window.confirm('Are you sure you want to delete this category?')) {
+            try {
+                const response = await fetch(`http://localhost:3001/api/admin/categories/${id}`, {
+                    method: 'DELETE',
+                    headers: {
+                        'Admin-Secret-Key': adminKey,
+                    },
+                });
 
-            if (response.ok) {
-                fetchCategories(); // Refresh the list
-            } else {
-                console.error('Failed to delete category');
+                if (response.ok) {
+                    fetchCategories();
+                    triggerCategoryChange();
+                } else {
+                    console.error('Failed to delete category');
+                }
+            } catch (error) {
+                console.error('Error deleting category:', error);
             }
-        } catch (error) {
-            console.error('Error deleting category:', error);
         }
     };
 
+
+
     return (
-        <div className="admin-container">
-            <h2 className="admin-title">Zarządzaj kategoriami</h2>
+        <div className="admin-container p-5">
+            <h2 className="admin-title text-2xl font-semibold mb-4">Zarządzaj kategoriami</h2>
             <CategoryForm
                 onSubmit={addOrUpdateCategory}
                 initialData={editingCategory || {}}
-                setEditingCategory={setEditingCategory} // Pass the function to CategoryForm
+                setEditingCategory={setEditingCategory}
             />
-            <ul className="admin-list">
+            <ul className="admin-list mt-4">
                 {categories.map((category) => (
-                    <li key={category._id} className="admin-list-item">
+                    <li key={category._id} className="admin-list-item bg-white p-4 rounded shadow mb-4">
                         {category.name}
-                        <div>
-                            <button className="edit" onClick={() => setEditingCategory(category)}>Edit</button>
-                            <button className="delete" onClick={() => deleteCategory(category._id)}>Delete</button>
+                        <div className="flex justify-between">
+                            <button className="edit bg-yellow-500 text-white font-bold uppercase text-sm px-3 py-1 rounded shadow hover:bg-yellow-600 transition ease-in-out duration-300" onClick={() => setEditingCategory(category)}>Edytuj</button>
+                            <button className="delete bg-red-500 text-white font-bold uppercase text-sm px-3 py-1 rounded shadow hover:bg-red-600 transition ease-in-out duration-300" onClick={() => deleteCategory(category._id)}>Usuń</button>
                         </div>
                     </li>
                 ))}
             </ul>
         </div>
-
     );
+
 };
 
 export default CategoriesManager;
